// src/models/interfaces/HistoryResponce.tsx
export default interface History {
  txtFaxID?: number;
  txtUID?: number;
  txtOrderNo?: string | number;
  txtDeviceId?: string;
  tokenid?: string;
  req_from?: string;
  txtNoOfPage?: number;
  isDelete?: number | boolean;
  txtCost?: string;
  txtcredits_per_page?: string;
  txtStatus?: string | undefined;
  txtRequestedAt?: Date | undefined;
  txtCompletedAt?: Date | undefined;
  txtTotalDuration?: string;
  txtPhone?: number;
  txtEmail?: string;
  txtName?: string;
  txtTotalCredit?: string | number;
  r_number?: string;
  r_status?: string;
  r_retry_count?: string;
  r_error_code?: null;
  r_error_type?: null;
  r_error_msg?: null;
  r_bitrate?: string;
  r_resolution?: string;
  txtCoverPage?: number;
  created?: Date;
  mobile?: null;
  images?: string[];
  txtAttachEmailToFax: string;
  txtCreditUpdate: number;
  txtDateAdded: Date;
  txtEmailNotification: string;
  txtExpDate: Date;
  txtFreeCredit: string;
  txtLoginType: boolean | null;
  txtOtp: number | null;
  txtOtpVerify: boolean | null;
  txtPassword: string | null;
  txtPlan: string | undefined;
  txtPushNotification: string | undefined;
  txtRole: string | null;
  txtSocialMediaId: boolean | null;
  txtPId: number;
  txtDate: Date;
  txtAmount: number;
  txtPaymentId: number | string;
  txtFreeTrial: number;
  txtCancelPlan: number;
}